small updates

+++ app/Http/Controllers/RoleController.php
@@ -12,11 +12,12 @@ class RoleController extends Controller
     /**
      * Display a listing of the resource.
      */
-    function __construct()    {
-         $this->middleware('permission:role-list|role-create|role-edit|role-delete', ['only' => ['index','store']]);
-         $this->middleware('permission:role-create', ['only' => ['create','store']]);
-         $this->middleware('permission:role-edit', ['only' => ['edit','update']]);
-         $this->middleware('permission:role-delete', ['only' => ['destroy']]);
+    function __construct()
+    {
+        $this->middleware('permission:role-list|role-create|role-edit|role-delete', ['only' => ['index', 'store']]);
+        $this->middleware('permission:role-create', ['only' => ['create', 'store']]);
+        $this->middleware('permission:role-edit', ['only' => ['edit', 'update']]);
+        $this->middleware('permission:role-delete', ['only' => ['destroy']]);
     }
 
     public function index()
@@ -45,9 +46,9 @@ class RoleController extends Controller
         ]);
         $role = Role::create(['name' => $request->input('name')]);
         $role->syncPermissions($request->input('permission'));
-    
+
         return redirect()->route('roles')
-                        ->with('success','Role created successfully');
+            ->with('success', 'Role created successfully');
     }
 
     /**
@@ -66,10 +67,10 @@ class RoleController extends Controller
         $role = Role::findOrFail(decrypt($id));
         $permissions = Permission::get();
         $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", decrypt($id))
-            ->pluck('role_has_permissions.permission_id','role_has_permissions.permission_id')
+            ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
             ->all();
-    
-        return view('backend.role.edit', compact('role','permissions','rolePermissions'));
+
+        return view('backend.role.edit', compact('role', 'permissions', 'rolePermissions'));
     }
 
     /**
@@ -78,26 +79,30 @@ class RoleController extends Controller
     public function update(Request $request, string $id)
     {
         $this->validate($request, [
-            'name' => 'required|unique:roles,name,'.$id,
+            'name' => 'required|unique:roles,name,' . $id,
             'permission' => 'required',
         ]);
-    
+
         $role = Role::findOrFail($id);
-        $role->name = $request->input('name');
+        $role->name = ($id == 1) ? 'Administrator' : $request->input('name');
         $role->save();
-    
+
         $role->syncPermissions($request->input('permission'));
-    
+
         return redirect()->route('roles')
-                        ->with('success','Role updated successfully');
+            ->with('success', 'Role updated successfully');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id){
-        Role::findOrFail(decrypt($id))->delete();
-        return redirect()->route('roles')
-                        ->with('success','Role deleted successfully');
+    public function destroy(string $id)
+    {
+        if (decrypt($id) != 1) :
+            Role::findOrFail(decrypt($id))->delete();
+            return redirect()->route('roles')->with('success', 'Role deleted successfully');
+        else :
+            return redirect()->route('roles')->with('error', 'Administrator role cannot be delete!');
+        endif;
     }
 }
